+++ packages/page-service/src/index.ts
@@ -1,7 +1,8 @@
 import * as path from 'path';
-import * as fsExtra from 'fs-extra';
+import * as fse from 'fs-extra';
 import * as prettier from 'prettier';
-import { IMaterialBlock } from '@iceworks/material-utils';
+import * as glob from 'glob';
+import { IMaterialBlock, IMaterialPage } from '@iceworks/material-utils';
 import {
   pagesPath,
   COMPONENT_DIR_NAME,
@@ -10,12 +11,19 @@ import {
   projectPath,
 } from '@iceworks/project-service';
 import { bulkGenerate } from '@iceworks/block-service';
+import {
+  bulkDownloadMaterials,
+  bulkInstallMaterialsDependencies,
+  getFolderLanguageType,
+} from '@iceworks/common-service';
 import * as upperCamelCase from 'uppercamelcase';
 import * as ejs from 'ejs';
+import * as transfromTsToJs from 'transform-ts-to-js';
 import reactPageTemplate from './templates/template.react';
 import vuePageTemplate from './templates/template.vue';
 import { bulkCreate } from './router';
 import i18n from './i18n';
+import renderEjsTemplates from './utils/renderEjsTemplates';
 
 export * from './router';
 
@@ -36,9 +44,9 @@ export const generate = async function ({
   const pagePath = path.join(pagesPath, pageName);
 
   // ensure that the root directory of the page store exists
-  await fsExtra.mkdirp(pagePath);
+  await fse.mkdirp(pagePath);
 
-  const isPagePathExists = await fsExtra.pathExists(pagePath);
+  const isPagePathExists = await fse.pathExists(pagePath);
   if (!isPagePathExists) {
     throw new Error(i18n.format('package.pageService.index.pagePathExistError', { name }));
   }
@@ -71,7 +79,7 @@ export const generate = async function ({
       parser: prettierParserType,
     });
 
-    await fsExtra.writeFile(dist, rendered, 'utf-8');
+    await fse.writeFile(dist, rendered, 'utf-8');
   } catch (error) {
     remove(pageName);
     throw error;
@@ -94,9 +102,72 @@ export async function createRouter(data) {
  * @param name {string} Page folder name
  */
 export const remove = async function (name: string) {
-  await fsExtra.remove(path.join(pagesPath, name));
+  await fse.remove(path.join(pagesPath, name));
 };
 
 export const addBlocks = async function (blocks: IMaterialBlock[], pageName: string) {
   return await bulkGenerate(blocks, path.join(pagesPath, pageName, COMPONENT_DIR_NAME));
+};
+
+export const getTemplateSchema = async (selectPage: IMaterialPage) => {
+  const templateTempDir = path.join(pagesPath, '.template');
+  try {
+    await bulkDownloadMaterials([selectPage], templateTempDir);
+    const templateSchema = await fse.readJSON(
+      path.join(pagesPath, '.template', selectPage.name, 'config', 'settings.json')
+    );
+    await fse.remove(templateTempDir);
+    return templateSchema;
+  } catch (err) {
+    await fse.remove(templateTempDir);
+    throw err;
+  }
+};
+
+export const createPage = async (selectPage: IMaterialPage) => {
+  const templateTempDir: string = path.join(pagesPath, '.template');
+  try {
+    await bulkDownloadMaterials([selectPage], templateTempDir);
+    await renderPage(selectPage);
+    await bulkInstallMaterialsDependencies([selectPage], projectPath);
+    await fse.remove(templateTempDir);
+  } catch (err) {
+    await fse.remove(templateTempDir);
+    throw err;
+  }
+};
+
+export const renderPage = async (page: IMaterialPage) => {
+  console.log('renderPage', page);
+  const templateName = page.name;
+  const pageName: string = upperCamelCase(page.pageName);
+  const templatePath: string = path.join(pagesPath, '.template', `${templateName}`);
+  const targetPath: string = path.join(pagesPath, `${pageName}`);
+  const templateData = page.templateData;
+
+  if (fse.existsSync(targetPath)) {
+    throw new Error(i18n.format('package.pageService.index.pagePathExistError', { name: pageName }));
+  }
+
+  await renderEjsTemplates(templateData, templatePath);
+  const pageSourceSrcPath = path.join(templatePath, 'src');
+  const pageType = getFolderLanguageType(pageSourceSrcPath);
+  const projectType = await getProjectLanguageType();
+
+  if (pageType === 'ts' && projectType === 'js') {
+    const files = glob.sync('**/*.@(ts|tsx)', {
+      cwd: pageSourceSrcPath,
+    });
+
+    console.log('transfrom ts to js', files.join(','));
+
+    transfromTsToJs(files, {
+      cwd: pageSourceSrcPath,
+      outDir: pageSourceSrcPath,
+      action: 'overwrite',
+    });
+  }
+  await fse.move(pageSourceSrcPath, targetPath);
+
+  return targetPath;
 };